Version 0.99 mplayer transparancy Lots of improvements. Bugfixes. etc.
